[feature] shell 재빌드 이슈 해결을 위한 dynamic class loading 기능 도입
