+++ shell_pkg/commands/__init__.py
@@ -1,0 +1,6 @@
+import os
+import sys
+
+current_dir = os.path.dirname(__file__)
+ROOT_dir = os.path.abspath(os.path.join(current_dir, '../..'))
+sys.path.append(ROOT_dir)
