+++ shell_pkg/commands/command.py
@@ -1,0 +1,29 @@
+import os
+import re
+from abc import ABC, abstractmethod
+from shell_pkg.constants import MIN_LBA, MAX_LBA
+
+from logger_pkg.logger import Logger
+
+
+class Command(ABC, Logger):
+    @abstractmethod
+    def execute(self):
+        pass
+
+    def is_invalid_lba(self, lba: str):
+        if type(lba) != str:
+            return True
+        if not lba.isdigit():
+            return True
+        if MIN_LBA > int(lba) or int(lba) > MAX_LBA:
+            return True
+        return False
+
+    def run_command(self, cmd):
+        os.system(cmd)
+
+    def is_invalid_value(self, value: str):
+        if type(value) != str:
+            return True
+        return not bool(re.match(r'^0x[0-9A-F]{8}$', value))
