+++ shell_pkg/shell.py
@@ -1,6 +1,7 @@
-from command import InvalidCommand, ExitCommand, HelpCommand, WriteCommand, ReadCommand, EraseCommand, \
-    EraseRangeCommand, FullreadCommand, \
-    FullwriteCommand, TestApp1Command, TestApp2Command
+import os
+import importlib.util
+from constants import COMMAND_DIR_PATH
+from commands.invalid_command import InvalidCommand
 
 
 class Shell:
@@ -15,31 +16,31 @@ class Shell:
             except Exception as e:
                 print(e)
 
+    def get_class_name(self, name: str):
+        components = name.split('_')
+        class_name = ''.join(x.title() for x in components)
+        return class_name
+
+    def get_command_module(self, args):
+        for filename in os.listdir(COMMAND_DIR_PATH):
+            if filename.endswith('.py'):
+                module_name = filename[:-3].split('_command')[0]
+                if module_name == args[0]:
+                    module_path = os.path.join(COMMAND_DIR_PATH, filename)
+                    spec = importlib.util.spec_from_file_location(module_name, module_path)
+                    module = importlib.util.module_from_spec(spec)
+                    spec.loader.exec_module(module)
+
+                    if hasattr(module, self.get_class_name(filename[:-3])):
+                        return getattr(module, self.get_class_name(filename[:-3]))(args)
+
+        return InvalidCommand()
+
     def determine_cmd(self, args):
         if len(args) == 0:
             return InvalidCommand()
-        elif args[0] == 'exit':
-            return ExitCommand(args)
-        elif args[0] == 'help':
-            return HelpCommand(args)
-        elif args[0] == 'write':
-            return WriteCommand(args)
-        elif args[0] == 'read':
-            return ReadCommand(args)
-        elif args[0] == 'erase':
-            return EraseCommand(args)
-        elif args[0] == 'erase_range':
-            return EraseRangeCommand(args)
-        elif args[0] == 'fullwrite':
-            return FullwriteCommand(args)
-        elif args[0] == 'fullread':
-            return FullreadCommand(args)
-        elif args[0] == 'testapp1':
-            return TestApp1Command(args)
-        elif args[0] == 'testapp2':
-            return TestApp2Command(args)
-        else:
-            return InvalidCommand()
+
+        return self.get_command_module(args)
 
 
 if __name__ == '__main__':
