+++ virtual_ssd_pkg/ssd.py
@@ -11,7 +11,8 @@ class VirtualSSD:
     def run(self):
         args = sys.argv[1:]
         cmd = self.command_factory.create_command(args)
-        cmd.execute(args)
+        cmd.execute()
+
 
 if __name__ == '__main__':
     ssd = VirtualSSD()
